+++ Packages/com.ibicha.youtube-player/Runtime/YoutubeDl/IYoutubeDl.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace YoutubePlayer
+{
+    interface IYoutubeDl
+    {
+        Task<T> GetVideoMetaDataAsync<T>(string youtubeUrl, YoutubeDlOptions options,
+            IEnumerable<string> schema, CancellationToken cancellationToken = default);
+    }
+}