+++ Packages/com.ibicha.youtube-player/CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.1] - 2021-10-10
+### Added
+- YoutubeDl will now try to use a local instance of youtube-dl instead of pinging a server. 
+  - This is turned on by default. Use `YoutubeDl.UseLocalInstance` to change the behaviour
+  - Playing the video will automatically download youtube-dl and check for updates.
+  - This feature is only available for Desktop platforms.
 
 ## [1.4.1] - 2021-07-02
 ### Changed